Add files via upload
This is your second assignment. The focus is on building some signal processing knowledge, and tools used to characterize the period of oscillatory (clean/noisy) data. The notebook focuses on the use of the scipy.signal.find_peaks function to quantify the signal period. At the end you will see some tasks that will help you explore the concepts of the notebook with more analytical focus, and present some limitations of this method as we move forward and start using the spectral methods (Wavelet and  Fourier Transforms).